[integ-tests] Remove test_create_wrong_pcluster_version
This test is intentionally using a custom AMI with the wrong ParallelCluster version and expecting error messages in cloud-init-output.log. This test is not stable because as soon as we make major cookbook changes, the error messages might not appear. This integration test was created by me. It really seems superfluous.

Signed-off-by: Hanwen <hanwenli@amazon.com>